move code to get is user in group to ActiveDirectorySearcher
